added dbinit class

+++ LibraryAPI/src/models/Library.java
@@ -1,6 +1,6 @@
 package models;
 import config.dbconfig;
-
+import config.dbinit;
 import java.util.*;
 import java.io.*;
 import java.sql.*;
@@ -13,15 +13,20 @@ public class Library{
 	private static String user = dbconfig.USER;
 	private static String password= dbconfig.PASSWORD;
     //***********************************
+	
 	private static Connection conn;
 	private static Statement stmt;
 	private static ResultSet rs;
-	private static String query;
 	
 	//currently logged in
-	private static String manager = ""; 
-	private static String associate = "";
-	private static String member = "";
+	private static String currentUser = "";
+	
+	//authority levels
+	//0: admin, managers
+	//1: associates
+	//2: members
+	//else: nonmembers
+	private static int authorityLevel = 3;
 	  
 	//create db connection  
 	public Library() throws ClassNotFoundException, SQLException{
@@ -29,118 +34,14 @@ public class Library{
 	    conn = DriverManager.getConnection(url, user, password);
 	    stmt = conn.createStatement();
 	    System.out.println("Database connected successfully");
-	    createDB();
-	}
-	
-	//create tables
-	private static void createDB() throws SQLException{
-	    try {
-	        stmt.executeUpdate("create schema if not exists Library;");
-	    } catch (SQLException ex) {
-	    }
-		createManagersTable();
-    	createAssociatesTable();
-	//    	createMembersTable();
-    	createBooksTable();
-    	createBookAuthorsTable();
-    	createBookKeywordsTable();
-    	
-    	System.out.println(getISBN("books_authors","Radric "));
-	}
-	
-	private static void createManagersTable() {
-		System.out.println("Creating Table: managers...");
-	    //init table
-	    String managersTable = "create table if not exists managers (" +
-	    					 	//"fname		varchar(15)		not null,"+
-	    					 	//"lname		varchar(15)		not null,"+
-	    					 	"username	varchar(15)		not null,"+
-	    					 	"password	varchar(15)		not null,"+
-	    					 	"loggedIn	boolean					,"+
-	    					 	"primary key(username));";
-	    try {
-	        stmt.executeUpdate(managersTable);
-	    } catch (SQLException ex) {
-	        System.out.println(ex.toString());
-	    }
-	}
-	
-	private static void createAssociatesTable() {
-		System.out.println("Creating Table: associates...");
-	    //init table
-	    String associatesTable = "create table if not exists associates (" +
-	    					 	//"fname		varchar(15)		not null,"+
-	    					 	//"lname		varchar(15)		not null,"+
-	    					 	"username	varchar(15)		not null,"+
-	    					 	"password	varchar(15)		not null,"+
-	    					 	"loggedIn	boolean					,"+
-	    					 	"primary key(username));";
-	    try {
-	        stmt.executeUpdate(associatesTable);
-	    } catch (SQLException ex) {
-	        System.out.println(ex.toString());
-	    }
-	}
-	
-	private static void createMembersTable() {
-		System.out.println("Creating Table: members...");
-	    //init table
-	    String membersTable = "create table if not exists members (" +
-	    					 	"fname						varchar(15)		not null,"+
-	    					 	"lname						varchar(15)		not null,"+
-	    					 	"address					varchar(50)		not null,"+
-	    					 	"phone						varchar(10)		not null,"+
-	    					 	"username					varchar(15)		not null,"+
-	    					 	"password					varchar(15)		not null,"+
-	    					 	"code						varchar(4)		not null,"+
-	    					 	"numBooksCheckedOut			int						,"+
-	    					 	"suspended					boolean					,"+
-	    					 	"loggedIn					boolean					,"+
-	    					 	"primary key(username));";
-	    try {
-	        stmt.executeUpdate(membersTable);
-	    } catch (SQLException ex) {
-	        System.out.println(ex.toString());
-	    }
-	}
-	
-	private static void createBooksTable() throws SQLException {
-		System.out.println("Creating Table: books...");
-		String booksTable = "create table if not exists books( "+
-							 "isbn				varchar(15)		not null, "+
-							 "name				varchar(250)	not null, "+
-							 "year				varchar(4)		not null, "+
-							 "availableCopies	int				not null, "+
-							 "holds				int				not null, "+
-							 "price				double				not null, "+
-				 			 "primary key(isbn))";
-		stmt.executeUpdate(booksTable);
-		loadBooks();
-	}
-	
-	private static void createBookKeywordsTable() throws SQLException {
-		System.out.println("Creating Table: books_keywords...");
-		String bookKeywords = 	"create table if not exists books_keywords( "+
-								"isbn		varchar(15)		not null, 		"+
-								"keyword	varchar(25)				,		"+
-								"primary key(isbn,keyword))";
-		stmt.executeUpdate(bookKeywords);
-		loadBookKeywords();
-				
-	}
-	
-	private static void createBookAuthorsTable() throws SQLException{
-		System.out.println("Creating Table: books_authors...");
-		String bookKeywords = 	"create table if not exists books_authors( "+
-								"isbn		varchar(15)		not null, 		"+
-								"author		varchar(25)				,		"+
-								"primary key(isbn,author))";
-		stmt.executeUpdate(bookKeywords);
-		loadBookAuthors();
+	    dbinit.createDB(stmt);
+	    printBooks();
 	}
 	
 	//functions to add tuples to the db
 	public static void createManager(String uname,String password) throws SQLException{
+		if (authorityLevel != 0) return;
+		
 		System.out.println("adding manager: " + uname);
 		String sql = 	"insert ignore into managers values (" +
 					 	"'"+uname+"',"+
@@ -151,6 +52,7 @@ public class Library{
 	}
 	
 	public static void createAssociate(String uname,String password) throws SQLException{
+		if (authorityLevel != 0) return;
 		System.out.println("adding associate: " + uname );
 		String sql = 	"insert ignore into associates values (" +
 					 	"'"+uname+"',"+
@@ -160,37 +62,9 @@ public class Library{
 		stmt.executeUpdate(sql);	
 	}
 	
-	private static void loadBookKeywords() throws SQLException {
-		System.out.println("populating table: books_keywords...");
-		File f = new File("src/config/books_keywords");
-		String sql = "LOAD DATA LOCAL INFILE '"+f.getAbsolutePath()+".txt' "+
-					 "INTO TABLE books_keywords "+
-					 "COLUMNS TERMINATED BY ',' "+
-					 "LINES STARTING BY '.'";
-		stmt.executeUpdate(sql);
-	}
-
-	private static void loadBooks() throws SQLException {
-		System.out.println("populating table: books...");
-		File f = new File("src/config/books");
-		String sql = "LOAD DATA LOCAL INFILE '"+f.getAbsolutePath()+".txt' "+
-					 "INTO TABLE books "+ 
-					 "COLUMNS TERMINATED BY ',' "+
-					 "LINES STARTING BY '.'";
-		stmt.executeUpdate(sql);
-	}
-	
-	private static void loadBookAuthors() throws SQLException {
-		System.out.println("populating table: books_authors...");
-		File f = new File("src/config/books_authors");
-		String sql = "LOAD DATA LOCAL INFILE '"+f.getAbsolutePath()+".txt' "+
-					 "INTO TABLE books_authors "+ 
-					 "COLUMNS TERMINATED BY ',' "+
-					 "LINES STARTING BY '.'";
-		stmt.executeUpdate(sql);
-	}
 	
 	public static void createBook(String isbn, String[] authors,String name, String year,int avail, double price, String[] keywords  ) throws SQLException{
+		if (authorityLevel != 0) return;
 		if (authors.length == 0)
 			return;
 		String sql = 	"insert ignore into books values( "+
@@ -222,9 +96,9 @@ public class Library{
 	//Remove Functions
 		//remove books by isbn
 	public static void removeBookISBN(String isbn) throws SQLException{
+		if (authorityLevel != 0) return;
 		//remove from books table if isbn is of required length
 		if(isbn.length() < 7) return;
-		
 		String sql = "delete from books "+
 					 "where ISBN LIKE %'"+isbn+"'%";
 		stmt.executeUpdate(sql);
@@ -239,17 +113,16 @@ public class Library{
 			  "where ISBN LIKE %'"+isbn+"'%";
 		stmt.executeUpdate(sql);
 	}
+
 		//remove by name
-	public static void removeBookNAME(String name){
-		
-	}
-		//remove by author
-	public static void removeBookAUTHOR(String name){
-		
+	
+	public static void removeBookNAME(String name) throws SQLException{
+		removeBookISBN(getISBN(name));
 	}
 	
 	//login functions
 		//manager
+	
 	public static void loginManager(String uname, String password) throws SQLException{
 	String sql =  	"SELECT m.username, m.password " +
 	          		"FROM managers m " +
@@ -265,12 +138,14 @@ public class Library{
 					"set loggedIn = 1 " +
 					"where username = '"+uname+"'";
 	    		stmt.executeUpdate(sql);
-	    		manager = uname;
+	    		currentUser = uname;
+	    		authorityLevel = 0;
 	    		Manager.handleMain();
 			}
 	}
 	
 		//associate
+	
 	public static void loginAssociate(String uname, String password) throws SQLException{
 	String sql =  	"SELECT a.username, a.password " +
 	          		"FROM associates a " +
@@ -286,36 +161,64 @@ public class Library{
 					"set loggedIn = 1 " +
 					"where username = '"+uname+"'";
 	    		stmt.executeUpdate(sql);
-	    		manager = uname;
+	    		currentUser = uname;
+	    		authorityLevel = 1;
 	    		Manager.handleMain();
 			}
 	}
 	
 	//logout functions
 		//manager
+	
 	public static void logoutManager() throws SQLException {
 		String sql = "update managers "+
-					"set loggedIn = 0 "+
-					"where username = '"+ manager +"'";
+					 "set loggedIn = 0 "+
+					 "where username = '"+ currentUser +"'";
 		stmt.executeUpdate(sql);
 	}
 	
 	//Search Functions
-	private static String getISBN(String table,String value) throws SQLException{
-		String sql=null;
-		if (table.equals("books_authors")){
-			sql = 	"select distinct isbn "+
-					"from "+table+" "+
-					"where author LIKE '%"+value+"%'";
-		}else{
-			sql = 	"select distinct isbn "+
-					"from "+table+" "+
-					"where name LIKE '%"+value+"%'";
-		}
-		
+		//get ISBN by name
+	
+	//Queries
+	private static String getISBN(String value) throws SQLException{
+		String sql = 	"select distinct isbn "+
+						"from books "+
+						"where name LIKE '%"+value+"%'";
 		rs = stmt.executeQuery(sql);
 		rs.next();
 		return rs.getString(1);				
 	}
 	
+	
+	//Display info
+	private static void printBooks() throws SQLException{
+		String sql =	"select distinct b.isbn, b.name,b.year, a.author,b.availableCopies "+
+					 	"from books b, books_authors a "+
+						"inner join books_authors "+
+					 	"where b.isbn = a.isbn";
+		rs = stmt.executeQuery(sql);
+		
+		System.out.println("----------------------------------------------------------------------------------------");
+		System.out.printf("|%-15s|%-33s|%-5s|%-22s\n","ISBN","Title (available)","Year","Author(s)");
+		System.out.println("----------------------------------------------------------------------------------------");
+		
+		String previsbn =null;
+		while (rs.next()){
+			if (rs.getString(1).equals(previsbn)){
+				System.out.print(", "+ rs.getString(4));
+				continue;
+			}else if (previsbn != null){
+				System.out.println("");
+			}
+			previsbn = rs.getString(1);
+			
+			System.out.printf("|%-15s|%-30s%3s|%-5s|"	,rs.getString(1)
+	   													,rs.getString(2).substring(0, Math.min(rs.getString(2).length(), 27))
+	   													,"("+rs.getString(5)+")"
+	   													,rs.getString(3));
+			System.out.print(rs.getString(4));
+		}
+		System.out.println("\n----------------------------------------------------------------------------------------");
+	}				 	
 }
